pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "moviepy>=2.1.2",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]
